Templates/python/database-enhanced/requirements.txt
mcp>=1.0.0
httpx>=0.25.0
keyring>=24.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
duckdb>=0.9.0 